Formatting
Updated to v1.0.10
Removed notification_illustration_mercenarycontractexpired.png as it was not required. We already have an atlas
Added try/catch so the game won't crash if GUI folder is missing. It still will work, just with pink textures
Added statistics log when leaving, replicating vanillas code
Updated workaround for when IsUnderMercenaryService be true