The game is not handling correctly MapEvent.AddInvolvedParty(bool notFromInit) correctly.
It's set everywhere to true, even when the sides have initialized the battle, not joined after it started.
I had to introduce a lot of transpiler patches to fix it. When the devs fix this, the transpilers will be in passive mode until deleted

Added an option to disable NPC's from using the entire system
Made some classes internal
Made Options more game like
Added a copy of the game's IMbEvent implementation to avoid calling Invoke via reflection